Implemented a remote server in the apptainer so that it doesnt have to load up each use
